# docs/swagger.yaml
basePath: /
definitions:
  handler.Contents:
    properties:
      movies:
        items:
          $ref: '#/definitions/models.Movie'
        type: array
      series:
        items:
          $ref: '#/definitions/models.Series'
        type: array
    type: object
  handler.Episode:
    properties:
      episode_link:
        type: string
    type: object
  handler.ErrorData:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
  handler.Season:
    properties:
      season_episodes:
        items:
          $ref: '#/definitions/handler.Episode'
        type: array
    type: object
  handler.Series:
    properties:
      series_age_category:
        type: string
      series_description:
        type: string
      series_director:
        type: string
      series_duration:
        type: integer
      series_genres:
        items:
          type: string
        type: array
      series_keywords:
        description: Consider an array for keywords
        items:
          type: string
        type: array
      series_producer:
        type: string
      series_seasons:
        items:
          $ref: '#/definitions/handler.Season'
        type: array
      series_title:
        type: string
      series_year:
        type: integer
    type: object
...
  handler.movieForm:
    properties:
      movie_age_category:
        type: string
      movie_description:
        type: string
      movie_director:
        type: string
      movie_duration:
        type: integer
      movie_genres:
        items:
          type: string
        type: array
      movie_keywords:
        items:
          type: string
        type: array
      movie_link:
        type: string
      movie_producer:
        type: string
      movie_title:
        type: string
      movie_year:
        type: integer
    type: object
  models.AgeCategory:
    properties:
      id:
        type: integer
      max_age:
        type: integer
      min_age:
        type: integer
    type: object
  models.Cover:
    properties:
      filename:
        type: string
      id:
        type: integer
      project_id:
        type: integer
    type: object
  models.Episode:
    properties:
      episode_number:
        type: integer
      id:
        type: integer
      link:
        type: string
      season_id:
        type: integer
    type: object
  models.Genre:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Keyword:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Movie:
    properties:
      age_categories:
        items:
          $ref: '#/definitions/models.AgeCategory'
        type: array
      cover:
        $ref: '#/definitions/models.Cover'
      description:
        type: string
      director:
        type: string
      duration:
        type: integer
      genres:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: integer
      keywords:
        items:
          $ref: '#/definitions/models.Keyword'
        type: array
      popularity:
        type: integer
      producer:
        type: string
      release_year:
        type: integer
      screenshots:
        items:
          $ref: '#/definitions/models.Screenshot'
        type: array
      title:
        type: string
      youtube_id:
        type: string
    type: object
  models.Project:
    properties:
      id:
        type: integer
      movies:
        items:
          $ref: '#/definitions/models.Movie'
        type: array
      project_id:
        type: integer
      project_type:
        type: string
      series:
        items:
          $ref: '#/definitions/models.Series'
        type: array
    type: object
  models.Screenshot:
    properties:
      filename:
        type: string
      id:
        type: integer
      project_id:
        type: integer
    type: object
  models.Season:
    properties:
      episodes:
        items:
          $ref: '#/definitions/models.Episode'
        type: array
      id:
        type: integer
      season_number:
        type: integer
      series_id:
        type: integer
    type: object
  models.Series:
    properties:
      age_categories:
        items:
          $ref: '#/definitions/models.AgeCategory'
        type: array
      cover:
        $ref: '#/definitions/models.Cover'
      description:
        type: string
      director:
        type: string
      duration:
        type: integer
      genres:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: integer
      keywords:
        items:
          $ref: '#/definitions/models.Keyword'
        type: array
      popularity:
        type: integer
      producer:
        type: string
      release_year:
        type: integer
      screenshots:
        items:
          $ref: '#/definitions/models.Screenshot'
        type: array
      seasons:
        items:
          $ref: '#/definitions/models.Season'
        type: array
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Type "JWT" token received from the server
  title: Ozinshe API
  version: "1.0"
paths:
...
  /favorites:
    get:
      description: Retrieves a list of all projects (movies and series) favorited
        by the current user.
      produces:
      - application/json
      responses:
        "200":
          description: List of favorited movies and series
          schema:
            items:
              $ref: '#/definitions/handler.Contents'
            type: array
        "400":
          description: Error getting favorites
          schema: {}
      security:
      - CookieAuth: []
      summary: Get a list of all favorited projects (movies and series) for the current
        user
      tags:
      - favorites
...
  /projects:
    get:
      description: Retrieves a list of all projects, including both movies and series.
      produces:
      - application/json
      responses:
        "200":
          description: List of movies and series
          schema:
            items:
              $ref: '#/definitions/handler.Contents'
            type: array
        "400":
          description: Error getting projects
          schema: {}
      summary: Get a list of all projects (movies and series)
      tags:
      - projects
  /projects/{id}:
    delete:
      description: Deletes an existing project based on the provided ID. Requires
        admin authorization.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project deleted successfully
        "400":
          description: Error deleting project
          schema: {}
      security:
      - CookieAuth: []
      summary: Delete an existing project (movie or series)
      tags:
      - projects
    get:
      description: Retrieves details of a project based on the provided ID.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project details (series)
          schema:
            $ref: '#/definitions/models.Series'
        "400":
          description: Error getting project
          schema: {}
      summary: Get details of a specific project (movie or series)
      tags:
      - projects
    put:
      consumes:
      - multipart/form-data
      description: Updates an existing project based on the provided ID and form data.
        Requires admin authorization.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project cover
        in: formData
        name: cover
        required: true
        type: file
      - description: Project screenshots
        in: formData
        name: screenshots
        required: true
        type: file
      - description: Project data (JSON)
        in: body
        name: movie_data
        required: true
        schema:
          $ref: '#/definitions/handler.movieForm'
      - description: Project data (JSON)
        in: body
        name: series_data
        required: true
        schema:
          $ref: '#/definitions/handler.Series'
      produces:
      - application/json
      responses:
        "200":
          description: Project updated successfully
        "400":
          description: Error updating project
          schema:
            $ref: '#/definitions/handler.ErrorData'
      security:
      - CookieAuth: []
      summary: Update an existing project (movie or series)
      tags:
      - projects
  /projects/{id}/favorites:
    delete:
      description: Allows the currently authenticated user to delete a project (movie
        or series) from their favorites list.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Favorite added successfully
        "400":
          description: Error adding to favorites
          schema: {}
      security:
      - CookieAuth: []
      summary: Delete a project from favorites
      tags:
      - favorites
    post:
      description: Allows the currently authenticated user to add a project (movie
        or series) to their favorites list.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Favorite added successfully
        "400":
          description: Error adding to favorites
          schema: {}
      security:
      - CookieAuth: []
      summary: Add a project to favorites
      tags:
      - favorites
  /projects/create-project:
    post:
      consumes:
      - multipart/form-data
      description: Creates a new project based on the provided project type and form
        data. Requires admin authorization.
      parameters:
      - description: Project type (movie or series)
        in: formData
        name: project_type
        required: true
        type: string
      - description: Project cover
        in: formData
        name: cover
        required: true
        type: file
      - description: Project screenshots
        in: formData
        name: screenshots
        required: true
        type: file
      - in: formData
        name: movie_age_category
        type: string
      - in: formData
        name: movie_description
        type: string
      - in: formData
        name: movie_director
        type: string
      - in: formData
        name: movie_duration
        type: integer
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: movie_genres
        type: array
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: movie_keywords
        type: array
      - in: formData
        name: movie_link
        type: string
      - in: formData
        name: movie_producer
        type: string
      - in: formData
        name: movie_title
        type: string
      - in: formData
        name: movie_year
        type: integer
      - in: formData
        name: series_age_category
        type: string
      - in: formData
        name: series_description
        type: string
      - in: formData
        name: series_director
        type: string
      - in: formData
        name: series_duration
        type: integer
      - collectionFormat: csv
        in: formData
        items:
          type: string
        name: series_genres
        type: array
      - collectionFormat: csv
        description: Consider an array for keywords
        in: formData
        items:
          type: string
        name: series_keywords
        type: array
      - in: formData
        name: series_producer
        type: string
      - in: formData
        name: series_title
        type: string
      - in: formData
        name: series_year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project created successfully
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Error creating project
          schema:
            $ref: '#/definitions/handler.ErrorData'
      security:
      - CookieAuthc: []
      summary: Create a new project (movie or series)
      tags:
      - projects
  /projects/search:
    get:
      description: Retrieves a list of projects (movies or series) matching the provided
        filter parameters.
      parameters:
      - description: Starting year for filter
        in: query
        name: year_start
        type: integer
      - description: Ending year for filter
        in: query
        name: year_end
        type: integer
      - description: Project type (movie or series)
        in: query
        name: project_type
        type: string
      - description: Popularity order (asc or desc)
        in: query
        name: popularity_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Filtered list of movies and series
          schema:
            items:
              $ref: '#/definitions/handler.Contents'
            type: array
        "400":
          description: Error getting filtered projects
          schema: {}
      summary: Get a filtered list of projects (movies or series) based on search
        criteria
      tags:
      - projects
...
swagger: "2.0"